+++ app/javascript/components/templates/GlobalNav.tsx
@@ -5,8 +5,8 @@ function GlobalNav() {
   return (
     <>
       <nav>
-        <Link to="/">Home</Link> | <Link to="games">Gomoku Narabe</Link>|{" "}
-        <Link to="login">Login</Link>
+        <Link to="/">Home</Link> | <Link to="posts">投稿一覧</Link>|{" "}
+        <Link to="login">Login</Link>| <Link to="users">ユーザー一覧</Link>
       </nav>
     </>
   );
