+++ app/javascript/components/templates/Login.tsx
@@ -1,0 +1,11 @@
+import React, { memo } from "react";
+// import SideBar from 'components/templates/SideBar'
+
+export const Login = memo(() => {
+  return (
+    <>
+      {/* <SideBar /> */}
+      <p>Loginページです</p>
+    </>
+  );
+});
