+++ app/javascript/components/templates/posts/PostAll.tsx
@@ -1,21 +1,21 @@
 import React from "react";
 
+import { useParams } from "react-router-dom";
+
 interface PostProps {
   title: string;
   body: string;
 }
-//下記記述が不要になった
-// HelloWorld.propTypes = {
-//   greeting: PropTypes.string
-// };
 
-function Post(props: PostProps) {
+function PostAll(props: PostProps) {
+  let { postId } = useParams();
   return (
     <React.Fragment>
       <h1>title: {props.title}</h1>
       <p>body: {props.body}</p>
+      <p>id: {postId}</p>
     </React.Fragment>
   );
 }
 
-export default Post;
+export default PostAll;
