+++ app/javascript/components/templates/posts/PostShow.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export function PostShow() {
+  return <div>Its UserAll page</div>;
+}
