+++ app/javascript/components/templates/users/UserShow.tsx
@@ -1,0 +1,5 @@
+import React, { memo } from "react";
+
+export function UserShow() {
+  return <div>Its UserAll page</div>;
+}
