+++ app/javascript/components/theme/theme.tsx
@@ -1,0 +1,14 @@
+// import { extendTheme } from "@chakra-ui/react";
+
+// const theme = extendTheme({
+//   styles: {
+//     global: {
+//       body: {
+//         backgroundColor: "red.100",
+//         color: "gray.800",
+//       },
+//     },
+//   },
+// });
+
+// export default theme;
