+++ config/routes.rb
@@ -13,5 +13,6 @@ Rails.application.routes.draw do
       resources :qiitas, only: %i[index]
     end
   end
+  get '*path', to: 'home#index'
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 end
