+++ package.json
@@ -10,6 +10,9 @@
   "dependencies": {
     "@babel/preset-react": "^7.17.12",
     "@babel/preset-typescript": "^7.17.12",
+    "@chakra-ui/react": "^2.2.4",
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
     "@rails/actioncable": "^6.0.0",
     "@rails/activestorage": "^6.0.0",
     "@rails/ujs": "^6.0.0",
@@ -17,6 +20,7 @@
     "@types/react": "^18.0.14",
     "@types/react-dom": "^18.0.5",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
+    "framer-motion": "^6.5.1",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
