+++ spec/models/post_spec.rb
@@ -6,4 +6,8 @@ RSpec.describe Post, type: :model do
   describe 'associations' do
     it { should belong_to(:user).class_name('User') }
   end
+
+  describe 'varidation' do
+    it { should validate_presence_of(:title) }
+  end
 end
